apprentissage couleur + nouvelle classe couleur

+++ src/general_methods/Color_Twister.java
@@ -1,9 +1,9 @@
 package general_methods;
 
 public class Color_Twister {
-	private short red;
-	private short green;
-	private short blue;
+	private int red;
+	private int green;
+	private int blue;
 	
 	private final short MIN = 0;
 	private final short MAX = 255;
@@ -11,26 +11,26 @@ public class Color_Twister {
 	 * Constructeur de la classe
 	 * Une couleur est composé de 3 valeurs numériques allant de 0 à 255 (rouge, vert, bleu)
 	 */
-	public Color_Twister(short r, short g, short b) {
+	public Color_Twister(int r, int g, int b) {
 		this.red = Limiteur(r);
 		this.green = Limiteur(g);
 		this.blue = Limiteur(b);
 	}
 	
 	// Getters & Setters
-	public short getRed() {
+	public int getRed() {
 		return red;
 	}
 	public void setRed(short red) {
 		this.red = Limiteur(red);
 	}
-	public short getGreen() {
+	public int getGreen() {
 		return green;
 	}
 	public void setGreen(short green) {
 		this.green = Limiteur(green);
 	}
-	public short getBlue() {
+	public int getBlue() {
 		return blue;
 	}
 	public void setBlue(short blue) {
@@ -43,7 +43,7 @@ public class Color_Twister {
 	 * @param a valeur d'une des 3 couleurs RGB
 	 * @return valeur comprise entre 0 et 255
 	 */
-	private short Limiteur(short a) {
+	private int Limiteur(int a) {
 		if(a < MIN) {return 0;}
 		else if (a > MAX) {return 255;}
 		else return a;
